requirements.txt
PyYAML~=6.0
chardet~=5.1.0